Exposed libxml2 interop
